+++ MME7/Gemfile
@@ -8,8 +8,8 @@ gem 'rails', '3.2.13'
 gem 'sqlite3'
 gem 'mail', '~> 2.5.3'
 gem 'carrierwave'
-gem 'rmagick'
-
+#gem 'rmagick'
+gem "rspec-rails", "~> 2.0"
 
 
 # Gems used only for assets and not required
